rpc客户端初始时调用服务端

+++ rpc-client/src/main/java/com/github/shrank/rpc/client/encoder/CalculateRequestEncoder.java
@@ -17,6 +17,6 @@ public class CalculateRequestEncoder extends MessageToByteEncoder<CalculateReque
         int a = calculateRequest.getA();
         int b = calculateRequest.getB();
         byteBuf.writeInt(a);
-        byteBuf.writeByte(b);
+        byteBuf.writeInt(b);
     }
 }
